--- .settings/org.eclipse.jdt.apt.core.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-org.eclipse.jdt.apt.aptEnabled=true
-org.eclipse.jdt.apt.genSrcDir=target\\generated-sources
