+++ src/main/java/ar/com/plug/examen/domain/model/entities/Client.java
@@ -1,6 +1,5 @@
-package ar.com.plug.examen.dao.entities;
+package ar.com.plug.examen.domain.model.entities;
 
-import java.util.List;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -9,20 +8,11 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-import ar.com.plug.examen.domain.model.Status;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Table(name="clients")
-@Getter
-@Setter
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
 public class Client {
 
 	@Id
