+++ src/main/java/ar/com/plug/examen/domain/model/entities/Transaction.java
@@ -1,4 +1,4 @@
-package ar.com.plug.examen.dao.entities;
+package ar.com.plug.examen.domain.model.entities;
 
 import java.util.List;
 
@@ -11,23 +11,14 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
 import ar.com.plug.examen.domain.model.Status;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Table(name="transactions")
-@Getter
-@Setter
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
 public class Transaction {
