+++ src/main/java/ar/com/plug/examen/domain/service/IProductService.java
@@ -2,7 +2,6 @@ package ar.com.plug.examen.domain.service;
 
 import java.util.List;
 
-import ar.com.plug.examen.dao.entities.Product;
 import ar.com.plug.examen.dto.ProductDto;
 
 public interface IProductService {
