+++ src/main/java/ar/com/plug/examen/dto/ClientDto.java
@@ -1,8 +1,6 @@
 package ar.com.plug.examen.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
