+++ src/test/java/ar/com/plug/examen/creator/TransactionCreator.java
@@ -3,10 +3,9 @@ package ar.com.plug.examen.creator;
 import java.util.ArrayList;
 import java.util.List;
 
-import ar.com.plug.examen.dao.entities.Client;
-import ar.com.plug.examen.dao.entities.Product;
-import ar.com.plug.examen.dao.entities.Transaction;
-import ar.com.plug.examen.domain.model.Status;
+import ar.com.plug.examen.domain.model.entities.Client;
+import ar.com.plug.examen.domain.model.entities.Product;
+import ar.com.plug.examen.domain.model.entities.Transaction;
 import ar.com.plug.examen.dto.TransactionDto;
 
 public class TransactionCreator {
