+++ src/test/java/ar/com/plug/examen/domain/service/controllerTest/ProductControllerTest.java
@@ -1,10 +1,10 @@
-package ar.com.plug.examen.domain.service;
+package ar.com.plug.examen.domain.service.controllerTest;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.mockito.ArgumentMatchers.any;
 
 import java.util.HashMap;
 import java.util.List;
@@ -22,51 +22,49 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import ar.com.plug.examen.Application;
-import ar.com.plug.examen.controller.ClientController;
-import ar.com.plug.examen.creator.ClientCreator;
-import ar.com.plug.examen.domain.service.impl.ClientServiceImpl;
-import ar.com.plug.examen.dto.ClientDto;
+import ar.com.plug.examen.app.rest.ProductController;
+import ar.com.plug.examen.creator.ProductCreator;
+import ar.com.plug.examen.domain.service.impl.ProductServiceImpl;
 import ar.com.plug.examen.dto.ProductDto;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ContextConfiguration(classes = Application.class)
-public class ClientControllerTest {
-
+public class ProductControllerTest {
+	
 	@InjectMocks
-	private ClientController controller;
+	private ProductController controller;
 	
 	@MockBean
-	private ClientServiceImpl service;
+	private ProductServiceImpl service;
 	
 	@Autowired
 	private TestRestTemplate restTemplate;
 	
-	private final String url = "/client/";
-	
-	
+	private final String url = "/product/";
+
 	@Test
-	public void createClientTest() {
-		ClientDto client = ClientCreator.createClientDtoMock();
+	public void createProductTest() {
+		ProductDto product = ProductCreator.createProductDto();
 		
-		ResponseEntity<ClientDto> response = restTemplate.postForEntity(url + "create", client, ClientDto.class);
+		ResponseEntity<ProductDto> response = restTemplate.postForEntity(url + "create", product, ProductDto.class);
 
 		assertEquals(HttpStatus.OK, response.getStatusCode());
 		
-		verify(this.service, times(1)).save(any(ClientDto.class));
+		verify(this.service, times(1)).save(any(ProductDto.class));
 		
 	}
 	
 	@Test
-	public void findAllClientTest() {
-		List<ClientDto> cList = ClientCreator
-				.createClientDtoList(ClientCreator.createClientDtoWithId(1L),
-						ClientCreator.createClientDtoWithId(2L));
+	public void findAllProductTest() {
+		List<ProductDto> pList = ProductCreator.createProductDtoList(
+				ProductCreator.createProductDtoWithId(1L),
+				ProductCreator.createProductDtoWithId(2L));
 		
-		when(service.findAll()).thenReturn(cList);
+		when(service.findAll()).thenReturn(pList);
 
 		ResponseEntity<List> response = restTemplate.getForEntity(url + "all", List.class);
 
@@ -77,10 +75,10 @@ public class ClientControllerTest {
 		
 	}
 	@Test
-	public void findClientByIdTest() throws Exception {
-		ClientDto client = ClientCreator.createClientDtoMock();
+	public void findProductByIdTest() throws Exception {
+		ProductDto product = ProductCreator.createProductDtoWithId(1L);
 		
-		when(service.findById(1L)).thenReturn(client);
+		when(service.findById(1L)).thenReturn(product);
 
 		ResponseEntity<ProductDto> response = restTemplate.getForEntity(url + 1L, ProductDto.class);
 
@@ -92,15 +90,15 @@ public class ClientControllerTest {
 	
 	
 	@Test
-	public void upgradeClientTest() throws Exception {
-		ClientDto client = ClientCreator.createClientDtoMock();
-		ClientDto cMock = ClientCreator.createClientDtoMock();
+	public void upgradeProductTest() throws Exception {
+		ProductDto product = ProductCreator.createProductDtoWithNameAndDescription("Product1","Product1");
+		ProductDto pMock = ProductCreator.createProductDtoWithId(1L);
 		
-		when(service.update(client)).thenReturn(cMock);
+		when(service.update(product)).thenReturn(pMock);
 		
-		HttpEntity<ClientDto> httpEntity = new HttpEntity<>(cMock);
+		HttpEntity<ProductDto> httpEntity = new HttpEntity<>(pMock);
 
-		ResponseEntity<ClientDto> response = restTemplate.exchange(url + "update", HttpMethod.PUT, httpEntity, ClientDto.class);
+		ResponseEntity<ProductDto> response = restTemplate.exchange(url + "update", HttpMethod.PUT, httpEntity, ProductDto.class);
 
 		assertEquals(HttpStatus.OK, response.getStatusCode());
 		
@@ -110,12 +108,11 @@ public class ClientControllerTest {
 	
 
 	@Test
-	public void deleteClientTest() {
+	public void deleteProductTest() {
 		Map<String, String> params = new HashMap<String, String>();
 	    params.put("id", "1");
 	    restTemplate.delete(url + "delete/{id}", params);
 
 	    verify(this.service, times(1)).delete(1L);
 	}
-	
 }
