+++ src/test/java/ar/com/plug/examen/domain/service/serviceTest/ClientServiceTest.java
@@ -1,4 +1,4 @@
-package ar.com.plug.examen.domain.service;
+package ar.com.plug.examen.domain.service.serviceTest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -16,8 +16,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import ar.com.plug.examen.creator.ClientCreator;
-import ar.com.plug.examen.dao.entities.Client;
-import ar.com.plug.examen.dao.jpa.ClientJpaDao;
+import ar.com.plug.examen.domain.model.entities.Client;
+import ar.com.plug.examen.domain.model.jpa.ClientJpaDao;
 import ar.com.plug.examen.domain.service.impl.ClientServiceImpl;
 import ar.com.plug.examen.dto.ClientDto;
 import ar.com.plug.examen.mapper.ClientMapper;
